+++ src/compare/max.ts
@@ -1,0 +1,11 @@
+import { bi2s, s2bi } from "../shared/utils.ts";
+import type { Max } from "./types.ts";
+import { maxInner } from "./utils.ts";
+
+/** This function returns max value. */
+export const max: Max = (array) => {
+  const arrayInner = array.map((str) => s2bi(str));
+  const [bi, fpe] = maxInner(arrayInner);
+
+  return bi2s(bi, fpe);
+};