+++ src/operations/add/tests/add.test.ts
@@ -1,5 +1,5 @@
 import { assertEquals } from "jsr:@std/assert";
-import { add } from "../main.ts";
+import { add } from "../../../../index.ts";
 
 Deno.test("12 + 21", () => {
   const res = add(["12", "21"]);