+++ src/operations/add/tests/addArray.test.ts
@@ -1,5 +1,5 @@
 import { assertEquals } from "jsr:@std/assert";
-import { add } from "../main.ts";
+import { add } from "../../../../index.ts";
 
 Deno.test("999 + 999 + 999", () => {
   const result = add(["999", "999", "999"]);