+++ src/operations/add/tests/random.test.ts
@@ -1,5 +1,5 @@
 import { assertEquals } from "jsr:@std/assert";
-import { add } from "../main.ts";
+import { add } from "../../../../index.ts";
 
 const randInRange = (min: number, max: number) => {
   return Math.floor(Math.random() * (max - min + 1) + min);